Changed tack with the CookieCleaner, and switched to only clean the cookie header when parsing it. Therefore then does not modify the request header that will be proxied onwards.
